// qldaFE/src/types/project.ts
export interface ProjectPhase {
  id?: string;
  name: string;
  description: string;
  order: number;
  status: "NOT_STARTED" | "IN_PROGRESS" | "COMPLETED" | "BLOCKED";
  startDate?: string;
  endDate?: string;
  documentProjectPhase?: ProjectPhaseDocument[];
  tasks: ProjectTask[];
  legalBasis?: string;
}

export interface ProjectPhaseCategory extends ProjectPhase {}

export interface ProjectTask {
  id: string;
  name: string;
  description: string;
  status: "NOT_STARTED" | "IN_PROGRESS" | "COMPLETED" | "BLOCKED";
  priority: "low" | "medium" | "high" | "critical";
  assignee?: string;
  startDate?: string;
  endDate?: string;
  progress: number;
  assignUser?: string;
  dependencies?: string[];
  documentsTask?: BaseDocument[];
  legalBasis?: string;
}

export interface BaseDocument {
  id: string;
  name: string;
  url?: string;
  uploadedAt: string;
  uploadedBy: string;
  type: FileExt;
}

export interface ProjectPhaseDocument extends BaseDocument {
  description?: string;
}